[optimise] Add Adam.
This commit adds a Java implementation of Adam, a gradient
descent algorithm. See the paper at [1] for details on the
implementation. On at least one test problem, this implementation
has been confirmed to produce identical output to the
implementation in recommendations.optimize.

Also, a record definition is added to cortex.optimise that acts as
a simple container for the AdamOptimizer Java class, and a Clojure
pure-function constructor has been added that allows configuration
of Adam's hyperparameters, with default values provided. The
constructor does not require a parameter count; this is inferred
at the first call of cortex.protocols/compute-parameters.

This implementation does not appear to work with examples/mnist.

[1]: http://arxiv.org/pdf/1412.6980v8.pdf