Add skeleton for new optimization architecture
Add several files:
- protocols, which defines the Function and Optimizer protocols
- functions, which provides implementations of the Function protocol
- optimizers, which provides implementations of the Optimizer protocol
- managers, which provides functions that perform gradient descent using
  Functions and Optimizers

To test out the new features, just mosey on over to
cortex.optimize.debug and run the two functions defined there.

The architecture introduced here is highly unstable and will likely
change significantly.

The biggest problem is currently that there is no obvious way to get the
Optimizer protocol to implement Associative so that internal state can
be extracted by the manager. We need to resolve this issue as soon as
possible, reworking the architecture if necessary.

Of course, the debug file will also be removed in due course.