Add get-state method to Optimizer protocol
Add a method to the Optimizer protocol for getting a map representing
the internal state of the optimizer (excluding the function value,
gradient, and next step, which should be dealt with by the manager).
This map can be lazy; we can implement something similar to [1] for
getting information lazily from the GPU back onto the CPU. Using [1]
directly, though, is undesirable because it has some odd behavior, such
as the vals function returning a sequence of Delays instead of a
sequence of values.

Update the Optimizer extend-protocol to implement the new method, and
print the state at each step in the default manager.

[1]: https://github.com/Malabarba/lazy-map-clojure