Use cortex protocols for new gradient descent fns
Delete all the new protocols, and port their implementations to instead
use the pre-existing cortex protocols (e.g. PParameters,
PGradientOptimiser, PGradient, PModule).

The protocols are extended only to APersistentMap, rather than
IPersistentMap, because all records are instances of IPersistentMap (but
not APersistentMap) and this caused the default implementation of
PGradient for various objects in cortex to be overridden, resulting in
failing tests.