Make certain incorrect optimisation calls throw earlier
It is very easy to forget that the optimisation functions sgd-clojure,
adadelta-clojure, etc. must be *called* in order to obtain an optimiser.
If you do forget, it is often hard to debug the error, since functions
can *also* be optimisers, but with a different interface. This commit
introduces a check to see if you have accidentally passed a function
which returns an optimiser map or function, and to throw an informative
error in that situation.