Import code en masse from think.recommend.optimize
This code represents most of think.recommend.optimize, as well as the
supporting functions from think.recommend.util. Much of it is adapted to
work with the new interface in cortex.optimise, or is refactored
entirely to improve clarity. A significant portion, though not all, of
it has been tested.