Allow specifying outputf params in vectors
Instead of just accepting functions (or function-like objects such as
keywords) as parameters for outputf, be smarter about parsing what comes
in. In particular, a vector is accepted, and will be interpreted as a
series of functions to apply to the state map, left to right. This
allows accessing deeply nested elements just by passing e.g. [:last
:params] instead of #(get-in % [:last :params]). Also, if the functions
are not actually functions, they will be interpreted as keys to look up
in the object resulting from previous transformations (as long as said
object is associative). This will allow passing e.g. [3 :last :params]
although I don't think this functionality is useful anywhere yet.