Make sformat more robust to %% and %n specifiers
The previous version of sformat had a number of minor bugs relating to
the use of the %% and %n specifiers (to create literal percent signs and
platform-independent newlines). This commit restructures the way in
which these specifiers are handled, viz. they are treated as normal
specifiers, but when sformat calls through to format the corresponding
argument is simply ignored. Also, an additional check is added to make
sure that %% and %n do not cause an argument to be skipped for the next
format specifier.