Add warning notice to six-hump-camel-back
The old version of check-gradient had a bug that prevented it from
checking any points with negative coodinates. Now that this bug was
fixed, this function no longer performs well on the gradient check,
implying that there is a flaw in its implementation for negative
numbers. Perhaps the derivative of a fractional power was taken
incorrectly?