[suite/train] Begin to separate concerns around `train-n`.
The notion of an `epoch-processor` is a good one. The problem is that the responsibilities around reporting and processing are spread around the entire training process. This is the first step (of likely many) to simplify high-level training by empowering the epoch processor to extensibly handle training regimes.

Training then can be pretty dumb, and a pluggable set of epoch processors can handle the demands of various training scenarios.