Removed PExecutionContext.
This commit removes a layer of abstraction that wasn't being used
anymore with the new compute infrastructure.  There was only one
implementation of the PExecutionContext, so it was removed and
integrated into the cortex.nn.execute namespace.  No doubt more moving
around to be done, but this as least simplifies things a little more.