Harold/ch773/decompose training mechanisms into a suite (#106)
* [suite/train/train-n] Rewrite this docstring.

* [suite/train] Begin to separate concerns around `train-n`.

The notion of an `epoch-processor` is a good one. The problem is that the responsibilities around reporting and processing are spread around the entire training process. This is the first step (of likely many) to simplify high-level training by empowering the epoch processor to extensibly handle training regimes.

Training then can be pretty dumb, and a pluggable set of epoch processors can handle the demands of various training scenarios.

* [suite/classification] Rename: best-network-function -> best-network-fn

For consistency and greppability.

* [suite/train] Removing unused function.

* [suite/train] A lying docstring and some trivial cleanups.