Training merge (#118)
* Initial steps towards decomposed training and inference.

Most of the changes here are minor things like renaming, formatting,
rearranging requires, adding comments and function docs.  The only real
addition is a run function that is the first take on a higher level API
to execute a model.  More on the way.

* Refactoring some names and adding docs.

* Renaming refactors...

* removed create prefix for most nouns to mirror Clojure
* changed getters that were using the a->b form to get-b
* fixed bugs that previous renaming created

* Removed PExecutionContext.

This commit removes a layer of abstraction that wasn't being used
anymore with the new compute infrastructure.  There was only one
implementation of the PExecutionContext, so it was removed and
integrated into the cortex.nn.execute namespace.  No doubt more moving
around to be done, but this as least simplifies things a little more.

* layer-graph->compute-graph renaming

* Minor changes.

* renamed :id->node-map to :nodes
* moved deep-merge to utils

* Streamlining namespaces around core abstractions.

* cortex.graph: ideally a pretty generic datastructure and algorithms
* cortex.layers: define core layers
* cortex.network: defining networks of layers and operating on them

* Fixing some includes that were broken in the last reorg.

Generally the goal has been to isolate the graph data structure and algorithm type stuff
in cortex.graph, and then have the neural net specific stuff in
cortex.layers and cortex.execute.

* working on the merge

* adding an all tests namespace

* Adding the previously ignored files.

* removing cuda from ignore

* [cuda/backend] Fix some merge errors, more tests passing now.

* [cuda/driver] Potential merge error here. How did this compile? Fewer errors but more assertion failures now.

* [execute] Merge error on getting output sizes.

* Concatenate test is passing.

* Unit tests are passing again.

* GPU->CPU fallback should be functional now.

* Getting importers to pass tests.