[verify/nn/train] Attempt to speed up mnist tests.
Re: https://github.com/thinktopic/cortex/issues/126

Recent refactors led to the mnist tests operating on a lot more data. This is not necessary to ensure the functionality these tests are trying to protect, and in the interest of speed operating on less data saves around 10min when running the tests on my local machine.