[dataset] Promote some code to util.
Some of the code in `datasets/util` was not really dataset specific and so can be safely moved to `cortex/util`.

What remains in `dataset.clj`, `datasets/mnist.clj`, and `datasets/util.clj` is now much closer to minimal, and represents good practices around utility datasets for testing and verification as well as basic information for users about datasets in general.